pyproject.toml
[tool.poetry]
name = "cinema"
version = "0.1.0"
description = ""
authors = ["frshd <frshdpnhi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
weasyprint = "^62.3"
django = "^5.1.1"
python-decouple = "^3.8"
boto3 = "^1.35.14"
django-storages = "^1.14.4"
djangorestframework = "^3.15.2"
psycopg = {extras = ["pool"], version = "^3.2.1"}
psycopg-binary = "^3.2.1"
ebooklib = "^0.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
